Time: 49 ms (6.67%), Space: 7.5 MB (13.33%) - LeetHub

+++ 1371-find-the-longest-substring-containing-vowels-in-even-counts/1371-find-the-longest-substring-containing-vowels-in-even-counts.go
@@ -1,7 +1,7 @@
 func findTheLongestSubstring(s string) int {
-  v := map[rune]int{ rune('a'): 1, rune('e'): 2, rune('i'): 4, rune('o'): 8, rune('u'): 16 }
-  m := map[int]int{ 0: -1 }
-  mask := 0
+  v := map[rune]uint8{ rune('a'): 1, rune('e'): 2, rune('i'): 4, rune('o'): 8, rune('u'): 16 }
+  m := map[uint8]int{ 0: -1 }
+  mask := uint8(0)
   ans := 0
   for i, r := range s {
     if bit, ok := v[r]; ok {
